requirements.txt
streamlit==1.41.1
pandas==2.2.3
numpy==2.2.3
plotly==6.0.0
yfinance==0.2.54
lxml==5.3.1
feedparser==6.0.11
python-dotenv==1.1.0
requests==2.32.3
pytest==8.3.4
